requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
msgraph-sdk==1.0.0
azure-identity==1.15.0
python-dotenv==1.0.0
mcp==1.21.2
ollama==0.6.1
requests==2.31.0
